Update sensor.py
The 2021.9/2021.10 update to HA.Core changes the energy class values with state class = STATE_CLASS_TOTAL_INCREASING to handle daily resets of production correct. For the Daily Energy it eliminates the need for the state class = STATE_CLASS_MEASUREMENT and the daily update of the reset time, which can be difficult to align with the start of production. on the daily production.